small typo

+++ ranker/include/ranker.h
@@ -18,8 +18,8 @@ struct url_score {
 
 class Ranker {
 public:
-const int static_weight = 0.1;
-const int dynamic_weight = 0.9;
+const float static_weight = 0.1;
+const float dynamic_weight = 0.9;
 ::vector<url_score *> getHighest(::vector<Match*>* matches, ISR* queryRoot);
 };
